feat: move rest of logic to internal
