+++ internal/analyzer/check/check.go
