+++ internal/analyzer/completion_item.go
