+++ internal/analyzer/package.go
