+++ internal/analyzer/scanner.go
