+++ internal/analyzer/scanner_test.go
