+++ internal/analyzer/sym_index.go
