+++ internal/analyzer/sym_index_test.go
