+++ internal/analyzer/testdata/bad.txt
