+++ internal/analyzer/testdata/pkg1.json
