+++ internal/analyzer/testdata/pkg2.json
